requirements.txt
configparser
enum34
libvirt-python
lockfile
lxml
paramiko
pbr
pyxdg
PyYAML
scp
setuptools
stevedore
future
wrapt
Jinja2
xmltodict
netaddr
